Removed Solr from Series Service
Removes all Solr functionality from the series service. Functions that
relied on the solr index were either removed or now query the database
directly.

- Moved the series filter query to the external api, while trying to
  preserve as much functionality as possible (for example, there is no
pagination, so instead an offset variable is offered)
