+++ modules/admin-ui/src/test/java/org/opencastproject/adminui/endpoint/TestSeriesEndpoint.java
@@ -46,7 +46,6 @@ import org.opencastproject.list.api.ResourceListQuery;
 import org.opencastproject.list.impl.ListProvidersServiceImpl;
 import org.opencastproject.metadata.dublincore.DublinCore;
 import org.opencastproject.metadata.dublincore.DublinCoreCatalog;
-import org.opencastproject.metadata.dublincore.DublinCoreCatalogList;
 import org.opencastproject.metadata.dublincore.DublinCoreXmlFormat;
 import org.opencastproject.metadata.dublincore.DublinCores;
 import org.opencastproject.metadata.dublincore.EncodingSchemeUtils;
@@ -62,7 +61,6 @@ import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.security.api.User;
 import org.opencastproject.security.api.UserDirectoryService;
 import org.opencastproject.series.api.SeriesException;
-import org.opencastproject.series.api.SeriesQuery;
 import org.opencastproject.series.api.SeriesService;
 import org.opencastproject.util.DateTimeSupport;
 import org.opencastproject.util.NotFoundException;
@@ -181,8 +179,6 @@ public class TestSeriesEndpoint extends SeriesEndpoint {
     final AccessControlList acl = new AccessControlList(
             new AccessControlEntry(anonymousRole, Permissions.Action.READ.toString(), true));
 
-    EasyMock.expect(seriesService.getSeries(EasyMock.anyObject(SeriesQuery.class)))
-            .andReturn(new DublinCoreCatalogList(catalogs, catalogs.size())).anyTimes();
     EasyMock.expect(seriesService.updateSeries(EasyMock.anyObject(DublinCoreCatalog.class))).andReturn(dc).anyTimes();
     EasyMock.expect(seriesService.getSeriesCount()).andReturn(3).anyTimes();
     EasyMock.expect(seriesService.getSeriesAccessControl(EasyMock.anyString())).andReturn(acl).anyTimes();
