+++ modules/series-service-api/src/main/java/org/opencastproject/series/api/SeriesService.java
@@ -22,7 +22,6 @@
 package org.opencastproject.series.api;
 
 import org.opencastproject.metadata.dublincore.DublinCoreCatalog;
-import org.opencastproject.metadata.dublincore.DublinCoreCatalogList;
 import org.opencastproject.security.api.AccessControlList;
 import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.util.NotFoundException;
@@ -137,30 +136,6 @@ public interface SeriesService {
    *           if exception occurred
    */
   AccessControlList getSeriesAccessControl(String seriesID) throws NotFoundException, SeriesException;
-
-  /**
-   * Search over series
-   *
-   * @param query
-   *          {@link SeriesQuery} representing query
-   * @return List of all matching series
-   * @throws UnauthorizedException
-   *           if the current user is not authorized to perform this action
-   * @throws SeriesException
-   *           if query could not be performed
-   */
-  DublinCoreCatalogList getSeries(SeriesQuery query) throws SeriesException, UnauthorizedException;
-
-  /**
-   * Returns a map of series Id to title of all series the user can access
-   *
-   * @return a map of series Id to title of all series the user can access
-   * @throws UnauthorizedException
-   *           if the current user is not authorized to perform this action
-   * @throws SeriesException
-   *           if query could not be performed
-   */
-  Map<String, String> getIdTitleMapOfAllSeries() throws SeriesException, UnauthorizedException;
 
   /**
    * Returns all the elements of a series in a map. The key of the map marks the element type. If
