--- modules/series-service-impl/src/main/java/org/opencastproject/series/impl/solr/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implementation of solr indexing for series.
- */
-package org.opencastproject.series.impl.solr;
