+++ modules/series-service-remote/src/main/java/org/opencastproject/series/remote/SeriesServiceRemoteImpl.java
@@ -31,46 +31,33 @@ import static org.apache.http.HttpStatus.SC_INTERNAL_SERVER_ERROR;
 import static org.apache.http.HttpStatus.SC_NOT_FOUND;
 import static org.apache.http.HttpStatus.SC_NO_CONTENT;
 import static org.apache.http.HttpStatus.SC_UNAUTHORIZED;
-import static org.opencastproject.util.doc.rest.RestParameter.Type.BOOLEAN;
-import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 
 import org.opencastproject.metadata.dublincore.DublinCore;
 import org.opencastproject.metadata.dublincore.DublinCoreCatalog;
-import org.opencastproject.metadata.dublincore.DublinCoreCatalogList;
 import org.opencastproject.metadata.dublincore.DublinCores;
 import org.opencastproject.security.api.AccessControlList;
 import org.opencastproject.security.api.AccessControlParser;
 import org.opencastproject.security.api.TrustedHttpClient;
 import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.series.api.SeriesException;
-import org.opencastproject.series.api.SeriesQuery;
 import org.opencastproject.series.api.SeriesService;
 import org.opencastproject.serviceregistry.api.RemoteBase;
 import org.opencastproject.serviceregistry.api.ServiceRegistry;
 import org.opencastproject.util.NotFoundException;
-import org.opencastproject.util.SolrUtils;
-import org.opencastproject.util.doc.rest.RestParameter;
-import org.opencastproject.util.doc.rest.RestQuery;
-import org.opencastproject.util.doc.rest.RestResponse;
 import org.opencastproject.util.doc.rest.RestService;
-
 
 import com.entwinemedia.fn.data.Opt;
 
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
 import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpPut;
-import org.apache.http.client.utils.URLEncodedUtils;
 import org.apache.http.entity.ByteArrayEntity;
 import org.apache.http.entity.ContentType;
 import org.apache.http.message.BasicNameValuePair;
-import org.apache.http.util.EntityUtils;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -81,7 +68,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
-import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -92,7 +78,6 @@ import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -132,9 +117,6 @@ public class SeriesServiceRemoteImpl extends RemoteBase implements SeriesService
   public SeriesServiceRemoteImpl() {
     super(JOB_TYPE);
   }
-
-  /** Default number of items on page */
-  private static final int DEFAULT_LIMIT = 20;
 
   /**
    * Sets the trusted http client
@@ -346,247 +328,6 @@ public class SeriesServiceRemoteImpl extends RemoteBase implements SeriesService
     throw new SeriesException("Unable to get series ACL from remote series index");
   }
 
-  @GET
-  @Produces(MediaType.APPLICATION_JSON)
-  @Path("series.json")
-  @RestQuery(
-      name = "listSeriesAsJson",
-      description = "Returns the series matching the query parameters",
-      returnDescription = "Returns the series search results as JSON",
-      restParameters = {
-          @RestParameter(
-              name = "q",
-              isRequired = false,
-              description = "Free text search",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "edit",
-              isRequired = false,
-              description = "Whether this query should return only series that are editable",
-              type = BOOLEAN
-          ),
-          @RestParameter(
-              name = "fuzzyMatch",
-              isRequired = false,
-              description = "Whether a partial match on series id is allowed, default is false",
-              type = BOOLEAN
-          ),
-          @RestParameter(
-              name = "seriesId",
-              isRequired = false,
-              description = "The series identifier",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "seriesTitle",
-              isRequired = false,
-              description = "The series title",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "creator",
-              isRequired = false,
-              description = "The series creator",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "contributor",
-              isRequired = false,
-              description = "The series contributor",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "publisher",
-              isRequired = false,
-              description = "The series publisher",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "rightsholder",
-              isRequired = false,
-              description = "The series rights holder",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "createdfrom",
-              isRequired = false,
-              description = "Filter results by created from (yyyy-MM-dd'T'HH:mm:ss'Z')",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "createdto",
-              isRequired = false,
-              description = "Filter results by created to (yyyy-MM-dd'T'HH:mm:ss'Z')",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "language",
-              isRequired = false,
-              description = "The series language",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "license",
-              isRequired = false,
-              description = "The series license",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "subject",
-              isRequired = false,
-              description = "The series subject",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "abstract",
-              isRequired = false,
-              description = "The series abstract",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "description",
-              isRequired = false,
-              description = "The series description",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "sort",
-              isRequired = false,
-              description = "The sort order.  May include any of the following: TITLE, SUBJECT, "
-                  + "CREATOR, PUBLISHER, CONTRIBUTOR, ABSTRACT, DESCRIPTION, CREATED, "
-                  + "AVAILABLE_FROM, AVAILABLE_TO, LANGUAGE, RIGHTS_HOLDER, SPATIAL, TEMPORAL, "
-                  + "IS_PART_OF, REPLACES, TYPE, ACCESS, LICENCE.  Add '_DESC' to reverse the "
-                  + "sort order (e.g. TITLE_DESC).",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "startPage",
-              isRequired = false,
-              description = "The page offset",
-              type = STRING
-          ),
-          @RestParameter(
-              name = "count",
-              isRequired = false,
-              description = "Results per page (max 100)",
-              type = STRING
-          )
-      },
-      responses = {
-          @RestResponse(
-              responseCode = SC_OK,
-              description = "The access control list."
-          ),
-          @RestResponse(
-              responseCode = SC_UNAUTHORIZED,
-              description = "If the current user is not authorized to perform this action"
-          )
-      }
-  )
-  public Response getSeriesAsJson(
-      @QueryParam("q") String text,
-      @QueryParam("seriesId") String seriesId,
-      @QueryParam("edit") Boolean edit,
-      @QueryParam("fuzzyMatch") Boolean fuzzyMatch,
-      @QueryParam("seriesTitle") String seriesTitle,
-      @QueryParam("creator") String creator,
-      @QueryParam("contributor") String contributor,
-      @QueryParam("publisher") String publisher,
-      @QueryParam("rightsholder") String rightsHolder,
-      @QueryParam("createdfrom") String createdFrom,
-      @QueryParam("createdto") String createdTo,
-      @QueryParam("language") String language,
-      @QueryParam("license") String license,
-      @QueryParam("subject") String subject,
-      @QueryParam("abstract") String seriesAbstract,
-      @QueryParam("description") String description,
-      @QueryParam("sort") String sort,
-      @QueryParam("startPage") String startPage,
-      @QueryParam("count") String count
-  ) throws UnauthorizedException {
-    try {
-      SeriesQuery seriesQuery = getSeries(
-          text, seriesId, edit, seriesTitle, creator, contributor, publisher,
-          rightsHolder, createdFrom, createdTo, language, license, subject, seriesAbstract, description, sort,
-          startPage, count, fuzzyMatch);
-      DublinCoreCatalogList result = getSeries(seriesQuery);
-      return Response.ok(result.getResultsAsJson()).build();
-    } catch (UnauthorizedException e) {
-      throw e;
-    } catch (Exception e) {
-      logger.warn("Could not perform search query: {}", e.getMessage());
-    }
-    throw new WebApplicationException(Response.Status.INTERNAL_SERVER_ERROR);
-  }
-
-  @Override
-  public DublinCoreCatalogList getSeries(SeriesQuery query) throws SeriesException, UnauthorizedException {
-    HttpGet get = new HttpGet(getSeriesUrl(query));
-    HttpResponse response = getResponse(get, SC_OK, SC_UNAUTHORIZED);
-    try {
-      if (response != null) {
-        int statusCode = response.getStatusLine().getStatusCode();
-        if (SC_OK == statusCode) {
-          DublinCoreCatalogList list = DublinCoreCatalogList.parse(
-              EntityUtils.toString(response.getEntity(), StandardCharsets.UTF_8));
-          logger.info("Successfully get series dublin core catalog list from the remote series index");
-          return list;
-        } else if (SC_UNAUTHORIZED == statusCode) {
-          throw new UnauthorizedException("Not authorized to get series from query");
-        }
-      }
-    } catch (UnauthorizedException e) {
-      throw e;
-    } catch (Exception e) {
-      throw new SeriesException("Unable to get series from query from remote series index: " + e);
-    } finally {
-      closeConnection(response);
-    }
-    throw new SeriesException("Unable to get series from query from remote series index: " + getSeriesUrl(query));
-  }
-
-  @Override
-  public Map<String, String> getIdTitleMapOfAllSeries() throws SeriesException, UnauthorizedException {
-    HttpGet get = new HttpGet("/allSeriesIdTitle.json");
-    HttpResponse response = getResponse(get, SC_OK, SC_UNAUTHORIZED, SC_INTERNAL_SERVER_ERROR);
-    try {
-      if (response != null) {
-        int statusCode = response.getStatusLine().getStatusCode();
-
-        if (SC_UNAUTHORIZED == statusCode) {
-          throw new UnauthorizedException("Not authorized to get series");
-        } else if (SC_OK == statusCode) {
-          String seriesJSON = EntityUtils.toString(response.getEntity(), StandardCharsets.UTF_8);
-          Object resultContainer = new JSONParser().parse(seriesJSON);
-          if (resultContainer instanceof JSONObject) {
-            Map<String, String> result = new HashMap<>();
-            JSONObject resultContainerJsonObj = (JSONObject) resultContainer;
-            JSONArray seriesJsonArr = resultContainerJsonObj.optJSONArray("series");
-            if (seriesJsonArr != null) {
-              for (int idx = 0; idx < seriesJsonArr.length(); idx++) {
-                JSONObject seriesJsonObj = seriesJsonArr.getJSONObject(idx);
-                String seriesId = seriesJsonObj.optString("identifier");
-                String seriesTitle = seriesJsonObj.optString("title");
-                if (StringUtils.isNotBlank(seriesId) && StringUtils.isNotEmpty(seriesTitle)) {
-                  result.put(seriesId, seriesTitle);
-                }
-              }
-            }
-            return result;
-          }
-        }
-      }
-    } catch (UnauthorizedException e) {
-      throw e;
-    } catch (Exception e) {
-      throw new SeriesException("Unable to get series from remote series index: " + e);
-    } finally {
-      closeConnection(response);
-    }
-    throw new SeriesException("Unable to get series from remote series index");
-  }
-
   @Override
   public int getSeriesCount() throws SeriesException {
     HttpGet get = new HttpGet("/count");
@@ -603,76 +344,6 @@ public class SeriesServiceRemoteImpl extends RemoteBase implements SeriesService
       closeConnection(response);
     }
     throw new SeriesException("Unable to count series from remote series index");
-  }
-
-  /**
-   * Builds the a series URL.
-   *
-   * @param q
-   *          the series query
-   * @return the series URL
-   */
-  private String getSeriesUrl(SeriesQuery q) {
-    StringBuilder url = new StringBuilder();
-    url.append("/series.xml?");
-
-    List<NameValuePair> queryStringParams = new ArrayList<>();
-    if (q.getText() != null) {
-      queryStringParams.add(new BasicNameValuePair("q", q.getText()));
-    }
-    if (q.getSeriesId() != null) {
-      queryStringParams.add(new BasicNameValuePair("seriesId", q.getSeriesId()));
-    }
-    queryStringParams.add(new BasicNameValuePair("edit", Boolean.toString(q.isEdit())));
-    queryStringParams.add(new BasicNameValuePair("fuzzyMatch", Boolean.toString(q.isFuzzyMatch())));
-    if (q.getSeriesTitle() != null) {
-      queryStringParams.add(new BasicNameValuePair("seriesTitle", q.getSeriesTitle()));
-    }
-    if (q.getCreator() != null) {
-      queryStringParams.add(new BasicNameValuePair("creator", q.getCreator()));
-    }
-    if (q.getContributor() != null) {
-      queryStringParams.add(new BasicNameValuePair("contributor", q.getContributor()));
-    }
-    if (q.getPublisher() != null) {
-      queryStringParams.add(new BasicNameValuePair("publisher", q.getPublisher()));
-    }
-    if (q.getRightsHolder() != null) {
-      queryStringParams.add(new BasicNameValuePair("rightsholder", q.getRightsHolder()));
-    }
-    if (q.getCreatedFrom() != null) {
-      queryStringParams.add(new BasicNameValuePair("createdfrom", SolrUtils.serializeDate(q.getCreatedFrom())));
-    }
-    if (q.getCreatedTo() != null) {
-      queryStringParams.add(new BasicNameValuePair("createdto", SolrUtils.serializeDate(q.getCreatedTo())));
-    }
-    if (q.getLanguage() != null) {
-      queryStringParams.add(new BasicNameValuePair("language", q.getLanguage()));
-    }
-    if (q.getLicense() != null) {
-      queryStringParams.add(new BasicNameValuePair("license", q.getLicense()));
-    }
-    if (q.getSubject() != null) {
-      queryStringParams.add(new BasicNameValuePair("subject", q.getSubject()));
-    }
-    if (q.getAbstract() != null) {
-      queryStringParams.add(new BasicNameValuePair("abstract", q.getAbstract()));
-    }
-    if (q.getDescription() != null) {
-      queryStringParams.add(new BasicNameValuePair("description", q.getDescription()));
-    }
-    if (q.getSort() != null) {
-      String sortString = q.getSort().toString();
-      if (!q.isSortAscending()) {
-        sortString = sortString.concat("_DESC");
-      }
-      queryStringParams.add(new BasicNameValuePair("sort", sortString));
-    }
-    queryStringParams.add(new BasicNameValuePair("startPage", Long.toString(q.getStartPage())));
-    queryStringParams.add(new BasicNameValuePair("count", Long.toString(q.getCount())));
-
-    url.append(URLEncodedUtils.format(queryStringParams, StandardCharsets.UTF_8));
-    return url.toString();
   }
 
   @Override
@@ -961,129 +632,5 @@ public class SeriesServiceRemoteImpl extends RemoteBase implements SeriesService
     } finally {
       closeConnection(response);
     }
-  }
-
-  private SeriesQuery getSeries(
-      String text,
-      String seriesId,
-      Boolean edit,
-      String seriesTitle,
-      String creator,
-      String contributor,
-      String publisher,
-      String rightsHolder,
-      String createdFrom,
-      String createdTo,
-      String language,
-      String license,
-      String subject,
-      String seriesAbstract,
-      String description,
-      String sort,
-      String startPageString,
-      String countString,
-      Boolean isFuzzyMatch
-  ) throws SeriesException, UnauthorizedException {
-    int startPage = 0;
-    if (StringUtils.isNotEmpty(startPageString)) {
-      try {
-        startPage = Integer.parseInt(startPageString);
-      } catch (NumberFormatException e) {
-        logger.warn("Bad start page parameter");
-      }
-      if (startPage < 0) {
-        startPage = 0;
-      }
-    }
-
-    int count = DEFAULT_LIMIT;
-    if (StringUtils.isNotEmpty(countString)) {
-      try {
-        count = Integer.parseInt(countString);
-      } catch (NumberFormatException e) {
-        logger.warn("Bad count parameter");
-      }
-      if (count < 1) {
-        count = DEFAULT_LIMIT;
-      }
-    }
-
-    SeriesQuery q = new SeriesQuery();
-    q.setCount(count);
-    q.setStartPage(startPage);
-    if (edit != null) {
-      q.setEdit(edit);
-    }
-    if (StringUtils.isNotEmpty(text)) {
-      q.setText(text);
-    }
-    if (StringUtils.isNotEmpty(seriesId)) {
-      q.setSeriesId(seriesId);
-    }
-    if (StringUtils.isNotEmpty(seriesTitle)) {
-      q.setSeriesTitle(seriesTitle);
-    }
-    if (StringUtils.isNotEmpty(creator)) {
-      q.setCreator(creator);
-    }
-    if (StringUtils.isNotEmpty(contributor)) {
-      q.setContributor(contributor);
-    }
-    if (StringUtils.isNotEmpty(language)) {
-      q.setLanguage(language);
-    }
-    if (StringUtils.isNotEmpty(license)) {
-      q.setLicense(license);
-    }
-    if (StringUtils.isNotEmpty(subject)) {
-      q.setSubject(subject);
-    }
-    if (StringUtils.isNotEmpty(publisher)) {
-      q.setPublisher(publisher);
-    }
-    if (StringUtils.isNotEmpty(seriesAbstract)) {
-      q.setSeriesAbstract(seriesAbstract);
-    }
-    if (StringUtils.isNotEmpty(description)) {
-      q.setDescription(description);
-    }
-    if (StringUtils.isNotEmpty(rightsHolder)) {
-      q.setRightsHolder(rightsHolder);
-    }
-    // allow seriesId wild card search
-    if (isFuzzyMatch != null) {
-      q.setFuzzyMatch(isFuzzyMatch.booleanValue());
-    }
-    try {
-      if (StringUtils.isNotEmpty(createdFrom)) {
-        q.setCreatedFrom(SolrUtils.parseDate(createdFrom));
-      }
-      if (StringUtils.isNotEmpty(createdTo)) {
-        q.setCreatedTo(SolrUtils.parseDate(createdTo));
-      }
-    } catch (ParseException e1) {
-      logger.warn("Could not parse date parameter: {}", e1);
-    }
-
-    if (StringUtils.isNotBlank(sort)) {
-      SeriesQuery.Sort sortField = null;
-      if (sort.endsWith("_DESC")) {
-        String enumKey = sort.substring(0, sort.length() - "_DESC".length()).toUpperCase();
-        try {
-          sortField = SeriesQuery.Sort.valueOf(enumKey);
-          q.withSort(sortField, false);
-        } catch (IllegalArgumentException e) {
-          logger.warn("No sort enum matches '{}'", enumKey);
-        }
-      } else {
-        try {
-          sortField = SeriesQuery.Sort.valueOf(sort);
-          q.withSort(sortField);
-        } catch (IllegalArgumentException e) {
-          logger.warn("No sort enum matches '{}'", sort);
-        }
-      }
-    }
-    return q;
   }
 }
